Replace complex class hierarchy for hbefa table loading with static loading function
